Following issue #29, this is a modified version of the Ant build file so that it produces a bundle that it is “OSGi consistent”, with a valid OSGi manifest. The modification consists in adding “by hand” the missing manifest properties.
The bundle is produced by a new target “osgi-dist” that generates a file
target/osgi/nrjavaserial-3.9.3.jar, which only differs from the
target/nrjavaserial-3.9.3.jar by the manifest.
This target is added in the default target’s dependencies.

Thera are also new properties in the build file:
  <property name="compile.target"      value="1.7"/>
  <property name="compile.source"      value="${compile.target}"/>
in fact it can be any target >= 1.4
and
  <property name="gnu.io.version" value="2.2"/>
which is used for the export package specification in the bundle’s manifest:
Export-Package: gnu.io;version="2.2"