Preserv errno from tcdrain for errror message.
When calling any function, the errno can be altered, so in order to be
sure that the errno printed is the actually interesting one, save it
between the point where you want to get the error kind and the output.
Without this, I got a lot of strange "java.io.IOException: No error in
nativeDrain" but in fact, there was a hidden error in tcdrain().