Fix IllegalMonitorStateException when thread closing port is interrupted
When a thread is closing the port it may not have obtained the writeLock when it is interrupted.
So it should only unlock the writeLock in the finally clause when the thread is holding that lock.
This fixes an IllegalMonitorStateException that would be thrown if the thread does not hold the lock.

Signed-off-by: Wouter Born <github@maindrain.net>