Add a read/write test, and a stub identifier test.
The read/write test expects to be configured (via environment variables;
see the `SerialPortExtension` Javadoc for details) with the names of two
serial ports connected to each other with a null modem cable. Armed with
this hardware, it verifies basic read/write functionality of the library
across a range of baud rates, and using both block and byte transfers.

The identifier test merely verifies a non-null response. Not terribly
useful at present, but I hope to use it to validate that the ports
configured for the other test are reported correctly (assuming that the
user configured them correctly, of course).