Refactor the native builds.
This reorganizes the makefile used to produce the native libraries. The
rules are now more product-oriented, and the rules for all target
platforms rely on the same core build/link targets. This makes it way
easier to see at a glance what the configuration is for a given platform,
and how it differs from any other platform. It also, IMHO, makes the
makefile much more readable, and the non-comment portion is much shorter.