Fix file deletion from make on Windows.
I forgot to check that the Windows file deletion rule worked. Turns out
my memory was foggy, and that the command is `RMDIR`, not `RMTREE`. (A
faulty and unnecessary portmanteau of `rmdir` and `deltree`, I suppose.)

I've also changed the name of the call from `rm` to `rmdir_r` to indicate
that it should only be used to operate on directories: unlike `rm -R`, `RMDIR`
cannot remove files directly (only those which exist under a directory which is
also being removed).