CI: use evergreen runners; update actions.
GitHub will continue to merrily deprecate and remove old runner
environments, as is their prerogative. However, the simplest way to
produce a library with a low minimum glibc version requirement is to
build in an environment having an old glibc. Not only does Ubuntu 18.04
have a fairly old glibc (v2.27), but the symbol versions our library
links against require only glibc v2.7 (on x86_64; somewhat higher for
some of the other architectures).

Having previously fixed cross-compilation compatibility with GCC 8, we
are no longer able to build with GCC 7-based compilers. Thankfully,
Ubuntu did package GCC 8 on 18.04; it's just not the default.