Linux ARM natives: add FP requirement to arch.
Without this change, compilation on newer versions of GCC fails:

    cc1: error: ‘-mfloat-abi=hard’: selected architecture lacks an FPU

The [GCC 8 ARM changes] note that “`-mfpu=auto` is now the default
setting unless the compiler has been configured with an explicit
`--with-fpu` option.” We were setting `-mfpu` to something other than
`auto`, but one possible interpretation of that sentence is “the value
of this option is now ignored unless you pass `--with-fpu`”. And of
course, didn't, so on architectures where an FPU is optional (ARMv6 and
ARMv7), we were being pushed onto a compile path that assumed the
absence.

A quick search uncovers a few other people dealing with this. Some
distro developers discussing [how to handle the switch] refer to GCC 8,
but the general consensus seems to be that the breaking change was
introduced [with GCC 11] in Debian/Ubuntu.

The GitHub Actions runners for Ubuntu 18.04 are now dead, but the last
[installed software list] indicates that they included GCC v7.5.0,
v9.4.0, and v10.3.0; presumably v7.5.0 was the default, which is why we
didn't run into this before now. (The logs and artifacts from the last
successful CI build are long expired.)

I expect that this breaks compilation on GCC <v8.

[GCC 8 ARM changes]: https://gcc.gnu.org/gcc-8/changes.html#arm
[installed software list]: https://github.com/actions/runner-images/blob/425daf97b4452130f0065e4fc58b5c8b34ab1941/images/linux/Ubuntu1804-Readme.md
[how to handle the switch]: https://gcc.gnu.org/pipermail/gcc/2021-September/237363.html
[with GCC 11]: https://bugs.launchpad.net/ubuntu/+source/gcc-defaults/+bug/1939379#yui_3_10_3_1_1692749691943_147